+++ CHANGES.md
@@ -7,15 +7,20 @@
   * add CLI binaries (to compile + upload) -> read from .adyenrc file
   * make compatible with Live system
 
-## v0.0.2
+## v0.0.4
+  * get the default path right
+  * add path to remote skins
+  * freeze local skins by default
+
+## v0.0.3
   * force update remote
   * .all unions remote + local (no more duplicates)
   * freeze locals to indicate missing remote endpoint
   * raise exception when page could not be loaded
 
 ## v0.0.2
- * cleanup
- * dep. fix
+  * cleanup
+  * dep. fix
 
 ## v0.0.1 - 10-04-2012
   _inital release_